Tidy up the previous attempts announcement as a component
